changed lilmachine to work with longs instead of ints

+++ src/day2/part1/Part1.java
@@ -8,19 +8,19 @@ import java.util.List;
 public class Part1 {
     public Part1(){ }
 
-    public List<Integer> computeProgram(List<Integer> i){
-        int programCounter = 0;
-        int opCode;
+    public List<Long> computeProgram(List<Long> i){
+        long programCounter = 0;
+        long opCode;
         do{
-            opCode = i.get(programCounter);
-            switch(opCode){
+            opCode = i.get((int) programCounter);
+            switch((int) opCode){
                 case 1: // add
-                    int addition = i.get(i.get(programCounter + 1)) + i.get(i.get(programCounter + 2));
-                    i.set(i.get(programCounter + 3), addition);
+                    long addition = i.get(Math.toIntExact(i.get((int) (programCounter + 1)))) + i.get(Math.toIntExact(i.get((int) (programCounter + 2))));
+                    i.set(Math.toIntExact(i.get((int) (programCounter + 3))), addition);
                     break;
                 case 2: // multiply
-                    int multiplication = i.get(i.get(programCounter + 1)) * i.get(i.get(programCounter + 2));
-                    i.set(i.get(programCounter + 3), multiplication);
+                    long multiplication = i.get(Math.toIntExact(i.get((int) (programCounter + 1)))) * i.get(Math.toIntExact(i.get((int) (programCounter + 2))));
+                    i.set(Math.toIntExact(i.get((int) (programCounter + 3))), multiplication);
                     break;
                 case 99: // halt
                     break;
@@ -33,7 +33,7 @@ public class Part1 {
 
     public static void main(String[] args) throws Exception{
         Part1 part1 = new Part1();
-        List<Integer> finalState = part1.computeProgram(
+        List<Long> finalState = part1.computeProgram(
                 new ProgramReader().readProgram("day2.txt")
         );
         System.out.println(finalState.get(0));
