+++ src/lilmachine/io/input/InputHandler.java
@@ -1,5 +1,5 @@
 package lilmachine.io.input;
 
 public interface InputHandler {
-    int getInput();
+    long getInput();
 }
