+++ src/lilmachine/io/output/SystemOutputHandler.java
@@ -2,7 +2,7 @@ package lilmachine.io.output;
 
 public class SystemOutputHandler implements OutputHandler{
     @Override
-    public void handleOutput(int output) {
+    public void handleOutput(long output) {
         System.out.println(output);
     }
 }
