+++ src/lilmachine/parameters/Parameter.java
@@ -6,26 +6,26 @@ import lilmachine.exceptions.IllegalModeException;
 public class Parameter {
 
     private ParameterMode mode;
-    private final int value;
+    private final long value;
 
-    public Parameter(int value, ParameterMode mode){
+    public Parameter(long value, ParameterMode mode){
         this.value = value;
         this.mode = mode;
     }
 
-    public int read(ProgramState state){
+    public long read(ProgramState state){
         return switch (mode) {
             case IMMEDIATE_MODE -> value;
-            case POSITION_MODE -> state.get(value);
-            case RELATIVE_MODE -> state.get(value + state.getRelativeBase());
+            case POSITION_MODE -> state.get((int) value);
+            case RELATIVE_MODE -> state.get((int) (value + state.getRelativeBase()));
         };
     }
 
-    public void write(ProgramState state, int data){
+    public void write(ProgramState state, long data){
         switch (mode) {
             case IMMEDIATE_MODE -> throw new IllegalModeException();
-            case POSITION_MODE -> state.set(value, data);
-            case RELATIVE_MODE -> state.set(value + state.getRelativeBase(), data);
+            case POSITION_MODE -> state.set((int) value, data);
+            case RELATIVE_MODE -> state.set((int) (value + state.getRelativeBase()), data);
         }
     }
